+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - **Setup flutter_localizations** - added `flutter_localizations` dependency
 - **Settings screen (UA)** - Settings screen and it's subscreens localization
 - **Convertors screen (UA)** - Convertors screen and all convertors subscreens localization
+- **Settings screen (UA)** - Settongs screen and Settings config screen localization
 
 
 ## [0.1.2] - 2026-04-26